refactor: extract socket message handlers (mirror the HTTP handler pattern)
The socket-server component inlined all four message handlers (request, recover,
outcome, request-validation) inside onConnection. Extract them into per-message
modules under logic/socket-server/handlers/, each a (context, data) => ackPayload
function — the socket analog of the HTTP handlers. getSocketRoutes() binds events
to handlers (like setupRouter), and the component wires them on each connection
with shared tracing / ack / error handling: a handler returns the payload to ack,
and an unexpected throw is reported to Sentry and acked as a generic error
(mirroring the HTTP errorHandler).

Behavior is unchanged — the socket integration tests (request/recover/outcome/
validation flows and their exact ack payloads) pass without modification.
