+++ src/logic/socket-server/routes.ts
@@ -1,0 +1,25 @@
+import { MessageType } from '../../ports/server/types'
+import { outcomeSocketHandler } from './handlers/outcome'
+import { recoverSocketHandler } from './handlers/recover'
+import { createRequestSocketHandler } from './handlers/request'
+import { requestValidationSocketHandler } from './handlers/request-validation'
+import { SocketRoute } from './types'
+
+export type SocketRoutesOptions = {
+  requestExpirationInSeconds: number
+  dclPersonalSignExpirationInSeconds: number
+}
+
+/**
+ * Binds each socket message type to its handler and tracing span — the socket analog of the HTTP
+ * `setupRouter`. The connection wrapper in the component consumes these to register `socket.on`
+ * listeners with shared span / ack / error handling.
+ */
+export function getSocketRoutes(options: SocketRoutesOptions): SocketRoute[] {
+  return [
+    { event: MessageType.REQUEST, span: 'websocket-request', handle: createRequestSocketHandler(options) },
+    { event: MessageType.RECOVER, span: 'websocket-recover', handle: recoverSocketHandler },
+    { event: MessageType.OUTCOME, span: 'websocket-outcome', handle: outcomeSocketHandler },
+    { event: MessageType.REQUEST_VALIDATION_STATUS, span: 'websocket-request-validation', handle: requestValidationSocketHandler }
+  ]
+}
